requirements.txt
# Core deep learning
torch==2.8.0
torchvision==0.23.0

# Graph neural networks
torch-geometric==2.6.1

# Open Graph Benchmark
ogb==1.3.6

# Numerical / scientific
numpy==2.0.2
scipy==1.13.1

# Machine learning utilities
scikit-learn==1.6.1

# Visualization
matplotlib==3.9.4
seaborn==0.13.2

# Experiment logging
tensorboard

# Config & utilities
PyYAML==6.0.3
tqdm==4.67.3
